fix: updates order-lineitem tests
