+++ server/products.js
@@ -36,7 +36,8 @@ router.route('/:id')
     }
   })
   .then(res => {
-    console.log(res.data)
+    console.log('res.data', res.data)
   })
   .catch(next)
 })
+
